ui changes and super refactor
